Move ExtensionHelper to main library, and move functionality for loading built-in extensions into main library instead of doing this separately for every API
