+++ benchmark/CMakeLists.txt
@@ -18,8 +18,6 @@ add_executable(benchmark_runner benchmark_runner.cpp interpreted_benchmark.cpp
 
 target_link_libraries(benchmark_runner duckdb imdb test_helpers)
 
-link_extension_libraries(benchmark_runner)
-
 if(${BUILD_TPCE})
   target_link_libraries(benchmark_runner tpce)
 endif()
