+++ benchmark/interpreted_benchmark.cpp
@@ -5,9 +5,9 @@
 #include <fstream>
 #include <sstream>
 #include "duckdb/main/query_profiler.hpp"
+#include "duckdb/main/extension_helper.hpp"
 #include "duckdb/common/string_util.hpp"
 #include "duckdb/main/client_context.hpp"
-#include "extension_helper.hpp"
 
 namespace duckdb {
