+++ extension/tpcds/CMakeLists.txt
@@ -7,4 +7,9 @@ include_directories(dsdgen/include)
 add_subdirectory(dsdgen)
 
 add_library(tpcds_extension STATIC tpcds-extension.cpp ${DSDGEN_OBJECT_FILES})
-# target_link_libraries(tpcds_extension dsdgen)
+
+install(
+  TARGETS tpcds_extension
+  EXPORT "${DUCKDB_EXPORT_SET}"
+  LIBRARY DESTINATION "${INSTALL_LIB_DIR}"
+  ARCHIVE DESTINATION "${INSTALL_LIB_DIR}")
