+++ src/CMakeLists.txt
@@ -10,23 +10,7 @@ if(NOT MSVC)
   )
 endif()
 
-# clang-tidy config from presto-cpp set(CMAKE_CXX_CLANG_TIDY clang-tidy
-# --checks=*,-abseil-*,-android-*,-cert-err58-cpp,-clang-analyzer-
-# osx-*,-cppcoreguidelines-avoid-c-arrays,-cppcoreguidelines-avoid-magic-
-# numbers,-cppcoreguidelines-pro-bounds-array-to-pointer-
-# decay,-cppcoreguidelines-pro-bounds-pointer-arithmetic,-cppcoreguidelines-pro-
-# type-reinterpret-cast,-cppcoreguidelines-pro-type-
-# vararg,-fuchsia-*,-google-*,-hicpp-avoid-c-arrays,-hicpp-deprecated-
-# headers,-hicpp-no-array-decay,-hicpp-use-equals-default,-hicpp-
-# vararg,-llvmlibc-*,-llvm-header-guard,-llvm-include-order,-mpi-*,-misc-non-
-# private-member-variables-in-classes,-misc-no-recursion,-misc-unused-
-# parameters,-modernize-avoid-c-arrays,-modernize-deprecated-headers,-modernize-
-# use-nodiscard,-modernize-use-trailing-return-
-# type,-objc-*,-openmp-*,-readability-avoid-const-params-in-decls,-readability-
-# convert-member-functions-to-static,-readability-magic-numbers,-zircon-*)
-
 if(AMALGAMATION_BUILD)
-
   if(WIN32)
     add_definitions(/bigobj)
   endif()
@@ -34,11 +18,13 @@ if(AMALGAMATION_BUILD)
   add_library(duckdb SHARED "${PROJECT_SOURCE_DIR}/src/amalgamation/duckdb.cpp")
   target_link_libraries(duckdb ${CMAKE_DL_LIBS})
   link_threads(duckdb)
+  link_extension_libraries(duckdb)
 
   add_library(duckdb_static STATIC
               "${PROJECT_SOURCE_DIR}/src/amalgamation/duckdb.cpp")
   target_link_libraries(duckdb_static ${CMAKE_DL_LIBS})
   link_threads(duckdb_static)
+  link_extension_libraries(duckdb_static)
 
   install(FILES "${PROJECT_SOURCE_DIR}/src/amalgamation/duckdb.hpp"
                 "${PROJECT_SOURCE_DIR}/src/include/duckdb.h"
@@ -74,10 +60,12 @@ else()
   add_library(duckdb SHARED ${ALL_OBJECT_FILES})
   target_link_libraries(duckdb ${DUCKDB_LINK_LIBS})
   link_threads(duckdb)
+  link_extension_libraries(duckdb)
 
   add_library(duckdb_static STATIC ${ALL_OBJECT_FILES})
   target_link_libraries(duckdb_static ${DUCKDB_LINK_LIBS})
   link_threads(duckdb_static)
+  link_extension_libraries(duckdb_static)
 
   target_include_directories(
     duckdb PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
