+++ src/include/duckdb/main/extension_helper.hpp
@@ -1,0 +1,26 @@
+//===----------------------------------------------------------------------===//
+//                         DuckDB
+//
+// duckdb/main/extension_helper.hpp
+//
+//
+//===----------------------------------------------------------------------===//
+
+#pragma once
+
+#include <string>
+#include "duckdb.hpp"
+
+namespace duckdb {
+class DuckDB;
+
+enum class ExtensionLoadResult : uint8_t { LOADED_EXTENSION = 0, EXTENSION_UNKNOWN = 1, NOT_LOADED = 2 };
+
+class ExtensionHelper {
+public:
+	static void LoadAllExtensions(DuckDB &db);
+
+	static ExtensionLoadResult LoadExtension(DuckDB &db, const std::string &extension);
+};
+
+} // namespace duckdb
