+++ test/CMakeLists.txt
@@ -41,8 +41,6 @@ else()
   target_link_libraries(unittest duckdb_static test_helpers)
 endif()
 
-link_extension_libraries(unittest)
-
 if(${BUILD_TPCH_EXTENSION})
   if(${BUILD_SQLSMITH})
     add_subdirectory(sqlsmith)
