+++ tools/rest/CMakeLists.txt
@@ -12,5 +12,3 @@ add_extension_definitions()
 
 target_link_libraries(duckdb_rest_server duckdb_static ${LINK_EXTRA})
 link_threads(duckdb_rest_server)
-
-link_extension_libraries(duckdb_rest_server)
